Make React context in the example app work with hot module replacement
