+++ examples/react-app/src/context/RpcContext.tsx
@@ -1,9 +1,6 @@
 import type { Rpc, RpcSubscriptions, SolanaRpcApiMainnet, SolanaRpcSubscriptionsApi } from '@solana/web3.js';
 import { createSolanaRpc, createSolanaRpcSubscriptions, devnet } from '@solana/web3.js';
-import type { ReactNode } from 'react';
-import { createContext, useContext, useMemo } from 'react';
-
-import { ChainContext } from './ChainContext';
+import { createContext } from 'react';
 
 export const RpcContext = createContext<{
     rpc: Rpc<SolanaRpcApiMainnet>; // Limit the API to only those methods found on Mainnet (ie. not `requestAirdrop`)
@@ -12,24 +9,3 @@ export const RpcContext = createContext<{
     rpc: createSolanaRpc(devnet('https://api.devnet.solana.com')),
     rpcSubscriptions: createSolanaRpcSubscriptions(devnet('wss://api.devnet.solana.com')),
 });
-
-type Props = Readonly<{
-    children: ReactNode;
-}>;
-
-export function RpcContextProvider({ children }: Props) {
-    const { solanaRpcSubscriptionsUrl, solanaRpcUrl } = useContext(ChainContext);
-    return (
-        <RpcContext.Provider
-            value={useMemo(
-                () => ({
-                    rpc: createSolanaRpc(solanaRpcUrl),
-                    rpcSubscriptions: createSolanaRpcSubscriptions(solanaRpcSubscriptionsUrl),
-                }),
-                [solanaRpcSubscriptionsUrl, solanaRpcUrl],
-            )}
-        >
-            {children}
-        </RpcContext.Provider>
-    );
-}
