+++ examples/react-app/src/main.tsx
@@ -6,9 +6,9 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 import { Nav } from './components/Nav.tsx';
-import { ChainContextProvider } from './context/ChainContext.tsx';
-import { RpcContextProvider } from './context/RpcContext.tsx';
-import { SelectedWalletAccountContextProvider } from './context/SelectedWalletAccountContext.tsx';
+import { ChainContextProvider } from './context/ChainContextProvider.tsx';
+import { RpcContextProvider } from './context/RpcContextProvider.tsx';
+import { SelectedWalletAccountContextProvider } from './context/SelectedWalletAccountContextProvider.tsx';
 import Root from './routes/root.tsx';
 
 const rootNode = document.getElementById('root')!;
